+++ src/cpart/cg.cpp
@@ -1,0 +1,16 @@
+//E函数,几何计算
+void E(){
+    hy("几何计算");
+    while(true){
+        line("-",20,"yellow",true);
+        print("输入任意键继续，输入“\033[1;33mc\033[0m”退出...","no",true);
+        cin >> q;
+        if (q == 'c' || q == 'C'){
+            clear();
+            break;
+        }
+        else{
+            //矩形 梯形 三角形 圆 面积,周长,体积
+        }
+    }
+}
